refactor: move call script back to core_utils
