+++ app/routers/setup.py
@@ -6,9 +6,8 @@ from pydantic import BaseModel
 
 from app.auth.auth_bearer import JWTBearer
 from app.auth.auth_handler import sign_jwt
-from app.core_utils import redis_get
+from app.core_utils import call_script, parse_key_value_lines, redis_get
 from app.repositories.system import name_valid, password_valid, shutdown
-from app.repositories.utils.raspiblitz import call_script, parse_key_value_lines
 
 router = APIRouter(prefix="/setup", tags=["Setup"])
